+++ src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std]
 
 pub mod pattern;
+pub mod postfx;
 pub mod state;